Update server tests
